volta list: refactor inventory module structure
- add modules for each collection implementation
- remove now-defunct (and previously misnamed) `serial` module
