+++ crates/volta-core/src/inventory/package.rs
@@ -1,0 +1,39 @@
+use std::collections::BTreeSet;
+
+use semver::Version;
+use volta_fail::Fallible;
+
+use crate::tool::{Package, PackageConfig};
+
+pub struct PackageCollection {
+    pub packages: BTreeSet<PackageConfig>,
+}
+
+impl PackageCollection {
+    // loads an empty PackageCollection
+    // ISSUE(#288) Collection only supports versions - for packages we also need names
+    pub(crate) fn load() -> Fallible<Self> {
+        Ok(PackageCollection {
+            packages: BTreeSet::new(),
+        })
+    }
+
+    pub(crate) fn contains(&self, name: &str, version: &Version) -> bool {
+        self.packages
+            .iter()
+            .find(|config| config.name == name && &config.version == version)
+            .is_some()
+    }
+}
+
+impl super::Collection for PackageCollection {
+    type Tool = Package;
+
+    fn add(&mut self, version: &Version) -> Fallible<()> {
+        unimplemented!()
+    }
+
+    fn remove(&mut self, version: &Version) -> Fallible<()> {
+        unimplemented!()
+    }
+}
